CV2-4072: add meComponent
