+++ src/app/components/source/UserInfo.js
@@ -20,6 +20,7 @@ const messages = defineMessages({
 
 const UserInfo = (props) => {
   if (props.user.source === null) return null;
+  const isUserSelf = (props.user.dbid === props.context.currentUser.dbid);
 
   return (
     <div className={styles['user-info-wrapper']}>
@@ -44,7 +45,7 @@ const UserInfo = (props) => {
                 size="default"
                 className="source__edit-source-button"
                 onClick={() => {
-                  if (props.user.dbid === props.context.currentUser.dbid) {
+                  if (isUserSelf) {
                     browserHistory.push('/check/me/edit');
                   } else {
                     browserHistory.push(`/check/user/${props.user.dbid}/edit`);
@@ -73,24 +74,27 @@ const UserInfo = (props) => {
             defaultMessage="{teamsCount, plural, one {# workspace} other {# workspaces}}"
             description="Count of how many work spaces this user account can access"
             values={{
-              teamsCount: props.user.team_users.edges.length || 0,
+              teamsCount: props.user.number_of_teams || 0,
             }}
           />
         </div>
-        <ButtonMain
-          className="user-menu__logout"
-          variant="contained"
-          theme="lightText"
-          size="default"
-          onClick={logout}
-          label={
-            <FormattedMessage
-              id="UserMenu.signOut"
-              defaultMessage="Sign Out"
-              description="This is the sign out button on the user profile page"
-            />
-          }
-        />
+        {
+          isUserSelf ?
+            <ButtonMain
+              className="user-menu__logout"
+              variant="contained"
+              theme="lightText"
+              size="default"
+              onClick={logout}
+              label={
+                <FormattedMessage
+                  id="UserMenu.signOut"
+                  defaultMessage="Sign Out"
+                  description="This is the sign out button on the user profile page"
+                />
+              }
+            /> : null
+        }
       </div>
     </div>
   );
